requirements.txt
requests
json
python-dotenv
langchain
openai
textblob
schedule
streamlit
altair
pandas